Added Comments to Associate

+++ Associate.cpp
@@ -12,8 +12,9 @@ using namespace std;
 int Associate::id_provider = 0;
 
 //Constructors
-Associate::Associate() : uniqueID(++id_provider), workingArea(nullptr) {}
-
+Associate::Associate() :
+		uniqueID(++id_provider), workingArea(nullptr) {
+}
 
 //Destructors
 Associate::~Associate() {
@@ -21,33 +22,32 @@ Associate::~Associate() {
 }
 
 //Get Methods
-string Associate::getWorkName() const{
+string Associate::getWorkName() const {
 	return this->workingArea->getName();
 }
 
-
-int Associate::getUniqueID() const{
+int Associate::getUniqueID() const {
 	return this->uniqueID;
 }
 
-string Associate::getStatus() const{
+string Associate::getStatus() const {
 	return this->status;
 }
 
-vector<int> Associate::getPaidYears() const{
+vector<int> Associate::getPaidYears() const {
 	return this->paidYears;
 }
 
-int Associate::getPersonalWallet() const{
+float Associate::getPersonalWallet() const {
 	return this->personalWallet;
 }
 
 //Set Methods
-void Associate::setWork(SubArea* sub){
+void Associate::setWork(SubArea* sub) {
 	this->workingArea = sub;
 }
 
-void Associate::setInterestAreas(vector <Area *> interest){
+void Associate::setInterestAreas(vector<Area *> interest) {
 	this->interestAreas = interest;
 }
 
@@ -55,7 +55,7 @@ void Associate::setStatus(string newStatus) {
 	this->status = newStatus;
 }
 
-void Associate::setPaidYears(vector<int> paid){
+void Associate::setPaidYears(vector<int> paid) {
 	this->paidYears = paid;
 }
 
@@ -63,21 +63,19 @@ void Associate::setPersonallWallet(int money) {
 	this->personalWallet = money;
 }
 
-
 //General Methods
 
-void Associate::addToWallet(int more_money){
+void Associate::addToWallet(int more_money) {
 	this->personalWallet += more_money;
 }
 
-void Associate::payFromWallet(int less_money){
+void Associate::payFromWallet(int less_money) {
 	this->personalWallet -= less_money;
 }
 
-void Associate::addPaidYear(int year){
+void Associate::addPaidYear(int year) {
 	this->paidYears.push_back(year);
 }
-
 
 void Associate::payYear(int year) {
 
@@ -99,7 +97,7 @@ void Associate::updateStatus() {
 
 	if (lastYearPaid >= Association::getCurrentYear() - 1) //if the last year he paid is the current last year or if he paid ahead
 		this->status = "contributor";
-	else if(lastYearPaid > (Association::getCurrentYear() - 5))
+	else if (lastYearPaid > (Association::getCurrentYear() - 5))
 		this->status = "subscriber";
 	else
 		this->status = "normal";
